@riversand963 2018-05-15
@trentxintong is this the result of profiling `db_stress`?