@trentxintong 2018-05-15
This is the result of TEST_TMPDIR=/dev/shm ./buck-out/gen/rocks/tools/rocks_db_bench --benchmarks=readwhilewriting --num=10000000 -benchmark_write_rate_limit=2000000 --threads=32