@siying 2018-05-17
The change makes sense, although I'll be very surprised that we can show any performance difference. These three are most likely immutable and very commonly used.