@trentxintong 2018-05-17
compare_ and head_ are very commonly used, allocator_ is not used as common. we can still have cache misses even they are immutable.