@siying 2018-05-17
@trentxintong I'm merging it as the benchmark shows no regression, and the new layout makes more sense.